+++ src/main/java/com/ktb7/pinpung/controller/PlaceController.java
@@ -2,12 +2,15 @@ package com.ktb7.pinpung.controller;
 
 import com.ktb7.pinpung.dto.PlaceInfoResponseDto;
 import com.ktb7.pinpung.dto.PlaceNearbyResponseDto;
+import com.ktb7.pinpung.dto.SearchResponseDto;
 import com.ktb7.pinpung.service.PlaceService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@Slf4j
 @RestController
 @RequestMapping("/places")
 public class PlaceController {
@@ -29,4 +32,15 @@ public class PlaceController {
         PlaceInfoResponseDto placeInfo = placeService.getPlaceInfo(placeId);
         return ResponseEntity.ok(placeInfo);
     }
+
+    @GetMapping("/tags-reviews")
+    public ResponseEntity<List<SearchResponseDto>> getPlacesWithReviewCountsAndTags(
+            @RequestParam List<Long> placeIds) {
+        if (placeIds.isEmpty()) {
+            return ResponseEntity.badRequest().build();
+        }
+        List<SearchResponseDto> places = placeService.getPlacesWithReviewCountsAndTags(placeIds);
+        return ResponseEntity.ok(places);
+    }
+
 }
